+++ pallets/charity/Cargo.toml
@@ -25,35 +25,35 @@ version = '1.1.0'
 [dependencies.frame-support]
 default_features = false
 git = 'https://github.com/paritytech/substrate.git'
-rev = '3e651110aa06aa835790df63410a29676243fc54'
+rev = '7c751574d8c53108e78440bfa5dffb80ab644972'
 
 [dependencies.frame-system]
 default_features = false
 git = 'https://github.com/paritytech/substrate.git'
-rev = '3e651110aa06aa835790df63410a29676243fc54'
+rev = '7c751574d8c53108e78440bfa5dffb80ab644972'
 
 [dependencies.balances]
 default_features = false
 git = 'https://github.com/paritytech/substrate.git'
 package = 'pallet-balances'
-rev = '3e651110aa06aa835790df63410a29676243fc54'
+rev = '7c751574d8c53108e78440bfa5dffb80ab644972'
 
 [dependencies.sp-runtime]
 default_features = false
 git = 'https://github.com/paritytech/substrate.git'
-rev = '3e651110aa06aa835790df63410a29676243fc54'
+rev = '7c751574d8c53108e78440bfa5dffb80ab644972'
 
 [dependencies.sp-std]
 default_features = false
 git = 'https://github.com/paritytech/substrate.git'
-rev = '3e651110aa06aa835790df63410a29676243fc54'
+rev = '7c751574d8c53108e78440bfa5dffb80ab644972'
 
 [dev-dependencies.sp-core]
 default_features = false
 git = 'https://github.com/paritytech/substrate.git'
-rev = '3e651110aa06aa835790df63410a29676243fc54'
+rev = '7c751574d8c53108e78440bfa5dffb80ab644972'
 
 [dev-dependencies.sp-io]
 default_features = false
 git = 'https://github.com/paritytech/substrate.git'
-rev = '3e651110aa06aa835790df63410a29676243fc54'
+rev = '7c751574d8c53108e78440bfa5dffb80ab644972'
